+++ packages/backend/src/models/DashboardModel/DashboardModel.mock.ts
@@ -174,6 +174,7 @@ export const dashboardWithVersionEntry: GetDashboardQuery = {
     first_name: 'firstName',
     last_name: 'lastName',
     pinned_list_uuid: 'pinnedUuid',
+    views: '1',
 };
 
 export const dashboardTileEntry: DashboardTileTable['base'] = {
@@ -272,6 +273,7 @@ export const expectedDashboard: Dashboard = {
         userUuid: 'userUuid',
     },
     pinnedListUuid: 'pinnedUuid',
+    views: 1,
 };
 
 export const expectedAllDashboards: DashboardBasicDetails[] = [
@@ -289,6 +291,7 @@ export const expectedAllDashboards: DashboardBasicDetails[] = [
         },
         spaceUuid: 'spaceUuid',
         pinnedListUuid: 'pinnedUuid',
+        views: 1,
     },
 ];