+++ packages/frontend/src/components/common/Dashboard/DashboardHeader.tsx
@@ -1,13 +1,8 @@
 import { Button, Classes, Divider, Intent, Menu } from '@blueprintjs/core';
 import { MenuItem2, Popover2, Tooltip2 } from '@blueprintjs/popover2';
-import {
-    Dashboard,
-    Space,
-    UpdateDashboardDetails,
-    UpdatedByUser,
-} from '@lightdash/common';
+import { Dashboard, Space, UpdatedByUser } from '@lightdash/common';
 import { useState } from 'react';
-import { Link, useHistory, useParams } from 'react-router-dom';
+import { useHistory, useParams } from 'react-router-dom';
 import { useApp } from '../../../providers/AppProvider';
 import { useTracking } from '../../../providers/TrackingProvider';
 import { EventName } from '../../../types/Events';
@@ -15,7 +10,6 @@ import AddTileButton from '../../DashboardTiles/AddTileButton';
 import ShareLinkButton from '../../ShareLinkButton';
 import DashboardUpdateModal from '../modal/DashboardUpdateModal';
 import {
-    IconWithRightMargin,
     PageActionsContainer,
     PageDetailsContainer,
     PageHeaderContainer,
@@ -24,8 +18,10 @@ import {
     PageTitleContainer,
     SeparatorDot,
 } from '../PageHeader';
+import SpaceInfo from '../PageHeader/SpaceInfo';
+import { UpdatedInfo } from '../PageHeader/UpdatedInfo';
+import ViewInfo from '../PageHeader/ViewInfo';
 import SpaceActionModal, { ActionType } from '../SpaceActionModal';
-import { UpdatedInfo } from '../UpdatedInfo';
 
 type DashboardHeaderProps = {
     spaces?: Space[];
@@ -34,6 +30,7 @@ type DashboardHeaderProps = {
     dashboardSpaceName?: string;
     dashboardSpaceUuid?: string;
     dashboardUpdatedAt: Date;
+    dashboardViews: number;
     dashboardUpdatedByUser?: UpdatedByUser;
     hasDashboardChanged: boolean;
     isEditMode: boolean;
@@ -52,6 +49,7 @@ const DashboardHeader = ({
     dashboardName,
     dashboardSpaceName,
     dashboardSpaceUuid,
+    dashboardViews,
     dashboardUpdatedAt,
     dashboardUpdatedByUser,
     hasDashboardChanged,
@@ -120,20 +118,18 @@ const DashboardHeader = ({
                         user={dashboardUpdatedByUser}
                     />
 
+                    <SeparatorDot icon="dot" size={6} />
+
+                    <ViewInfo views={dashboardViews} />
+
                     {dashboardSpaceName && (
                         <>
                             <SeparatorDot icon="dot" size={6} />
 
-                            <IconWithRightMargin
-                                icon="folder-close"
-                                size={10}
+                            <SpaceInfo
+                                link={`/projects/${projectUuid}/spaces/${dashboardSpaceUuid}`}
+                                name={dashboardSpaceName}
                             />
-
-                            <Link
-                                to={`/projects/${projectUuid}/spaces/${dashboardSpaceUuid}`}
-                            >
-                                {dashboardSpaceName}
-                            </Link>
                         </>
                     )}
                 </PageDetailsContainer>